Adressed PR feedback
New: merged userIcon and externalicon into new customIcons
