+++ apps/OpenSpace/ext/launcher/src/customicons.cpp
@@ -22,12 +22,28 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
  ****************************************************************************************/
 
-#include "externalicon.h"
+#include "customicons.h"
 
 #include <QColor>
 #include <QFont>
 #include <QPainter>
 #include <QPixmap>
+
+QIcon userIcon() {
+    QPixmap px = QPixmap(40, 50);
+    px.fill(Qt::transparent);
+
+    QPainter painter = QPainter(&px);
+    painter.setBrush(QColor(183, 211, 149, 255));
+    painter.drawEllipse(0, 10, 38, 38);
+
+    QFont f = QFont("Arial");
+    f.setPixelSize(28);
+    f.setBold(true);
+    painter.setFont(f);
+    painter.drawText(0, 10, 40, 40, Qt::AlignCenter, "U");
+    return QIcon(px);
+}
 
 QIcon externalIcon() {
     QPixmap px = QPixmap(40, 50);
